moved reset monitoring to monitoring page
