+++ dynamic-mapping-ui/src/monitoring/grid/monitoring.component.html
@@ -21,6 +21,15 @@
 
 <c8y-title translate>Mapping monitoring</c8y-title>
 <c8y-action-bar-item [placement]="'right'">
+	<button
+    class="btn btn-link"
+    title="{{ 'Reset monitoring' | translate }}"
+    (click)="resetStatusMapping()"
+    [disabled]="!feature?.userHasMappingAdminRole"
+  >
+    <i c8yIcon="reset"></i>
+    {{ 'Reset monitoring' | translate }}
+  </button>
   <button
     class="btn btn-link"
     title="{{ 'Refresh' | translate }}"
